Workaround for s3fs filesystem
This script doesn't work very well is you have your files stored in an
S3 bucker mounted as a directory with s3fs.  This is because s3fs is
not very good at metadata operations, like listing files and
directories. The script currently loops through each file in the
directory.  As a workaround in this case, you can use sync-glacier2.py,
which relies on the database to get metadata instead of the filesystem.